+++ models/genre_model.py
@@ -10,7 +10,7 @@ class Genre(db.Model):
     __tablename__ = 'genres'
 
     # primary key
-    genre_id = db.Column(db.Integer(128), primary_key=True)
+    genre_id = db.Column(db.Integer, primary_key=True)
 
     # relationships
     # genre_relations
